feat(AAPD-825): no need for pass through index
