+++ packages/web-comment/src/routes/appeal-open-comment/index.js
@@ -1,3 +1,9 @@
-module.exports = {
-	...require('./router')
-};
+const express = require('express');
+
+const router = express.Router();
+
+router.get('/', (req, res) => {
+	res.render(`appeal-open-comment/index`);
+});
+
+module.exports = { router };
