--- packages/web-comment/src/routes/appeal-open-comment/router.js
@@ -1,9 +1,0 @@
-const express = require('express');
-
-const router = express.Router();
-
-router.get('/', (req, res) => {
-	res.render(`appeal-open-comment/index`);
-});
-
-module.exports = { router };
